feat : 상품 상세 페이지 채팅, 상품별 채팅 분류 / fix : 일반 게시판 채팅 삭제

+++ src/main/java/com/example/demo/domain/chat/model/ChatRoom.java
@@ -45,7 +45,7 @@ import java.util.Date;
 @AllArgsConstructor
 public class ChatRoom {
     private Long chatRoomId;
-    private Long postId;
+    private int postId;
     private Integer userId;
     private String isGroupChat;
     private Date createdAt;
